Date Range & Fixes
Ability to input a date range. Will always default to returning an
array with the same date, but some situations may result in a range.
This build is not entirely tested yet. Also, this fixes some bugs
including an issue with multi-digit times and adds the ability to
understand “midnight” and “noon” then translate it to a date object
correctly.